requirements.txt
# Core dependencies for video processing
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0

# For matplotlib-based visualizations
matplotlib>=3.7.0

# Torch is typically provided by ComfyUI environment
# torch>=2.0.0
# torchvision>=0.15.0

# Optional: For ANSI color support on Windows
colorama>=0.4.6

# For lip sync / phoneme recognition
vosk>=0.3.45

# For BiRefNet background removal (Hugging Face transformers)
transformers>=4.36.0

# For FAL AI API nodes
fal-client>=0.4.0
requests>=2.28.0

# For MatAnyone video matting (YAML config loading)
omegaconf>=2.3.0

# For CorridorKey neural green screen keying (Hiera backbone)
timm>=1.0.0
einops>=0.8.0

# For PSD layer decomposition (PSD Layer Splitter node)
psd-tools>=1.9.0
